services/resource-matcher/requirements.txt
fastapi>=0.95.0
uvicorn[standard]>=0.22.0
databases[postgresql]>=0.7.0
asyncpg>=0.27.0
python-dotenv>=1.0.0
boto3>=1.30.0
psycopg2-binary>=2.9.6
pgvector>=0.3.0
anthropic>=0.12.0
